+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   no converters or converters which result in a [collections.abc.Sized][] compatible value.
 - Support for specifying the length of a string argument in annotation command declaration through
   [tanjun.annotations.Length][].
+- [AutocompleteContext.triggering_name][tanjun.context.autocomplete.AutocompleteContext.triggering_name]
+  which acts like [SlashContext.triggering_name][tanjun.context.slash.SlashContext.triggering_name].
 
 ### Changed
 - Bumped the minimum Hikari version to `2.0.0.dev111`.
@@ -18,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Make [tanjun.commands.slash.SlashCommandGroup.as_sub_command][]'s typing more
   flexible to allow decorating other command objects.
+- [SlashContext.triggering_name][tanjun.context.slash.SlashContext.triggering_name] now returns
+  the full triggering command name for sub-commands instead of just the top level command's name.
 
 ## [2.7.0a1] - 2022-09-24
 ### Added